backend/requirements.txt
fastapi
openai
qdrant-client
fastembed
python-dotenv
uvicorn
